requirements.txt
pyfiglet==0.8.post1
pypng==0.0.20   
PyQRCode==1.2.1
